Changed to download instead of clone.
The git directory (the directory with the name .git) is not needed.

Changed to specify user, repository, and branch which is hard to forget.

Changed to explicitly return the exit status, and return 1 if error.

Changed to use "&&" operator instead of if statement when executing one.

Changed the error message to be displayed in red.