Organized zone output.
Do not output if it is a samba zone.

Output only when the combination of view and suffix is correct.

The zone name is the correct (or the removed if it ends in -external).

Change the path to configuration depending on whether dynamic or not.

If external and secondary DNS is specified, output about zone transfer.